requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url
#
asgiref==3.3.4
    # via django
asn1crypto==1.4.0
    # via -r requirements.in
babel==2.9.1
    # via -r requirements.in
beautifulsoup4==4.9.3
    # via -r requirements.in
bleach==3.3.0
    # via django-fobi
boto3==1.17.73
    # via -r requirements.in
boto==2.49.0
    # via -r requirements.in
botocore==1.20.73
    # via
    #   boto3
    #   s3transfer
brotlipy==0.7.0
    # via -r requirements.in
cachetools==4.2.2
    # via google-auth
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via
    #   brotlipy
    #   cryptography
chardet==4.0.0
    # via requests
cryptography==3.4.7
    # via pyopenssl
dj-database-url==0.5.0
    # via -r requirements.in
django-admin-sortable2==1.0
    # via -r requirements.in
django-appconf==1.0.4
    # via
    #   django-compressor
    #   django-image-cropping
django-autoslug-iplweb==1.9.5
    # via -r requirements.in
django-autoslug==1.9.8
    # via django-fobi
django-ckeditor==6.0.0
    # via -r requirements.in
django-compressor==2.4.1
    # via -r requirements.in
django-crispy-forms==1.11.2
    # via -r requirements.in
django-debug-toolbar==3.2.1
    # via -r requirements.in
django-dynamic-forms==0.5.3
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-extra-views==0.13.0
    # via -r requirements.in
django-fobi==0.17.1
    # via -r requirements.in
django-formtools==2.3
    # via django-fobi
django-haystack==3.0
    # via -r requirements.in
django-image-cropping==1.5.0
    # via -r requirements.in
django-js-asset==1.2.2
    # via
    #   django-ckeditor
    #   django-mptt
django-mptt==0.12.0
    # via django-polymorphic-tree
django-nine==0.2.4
    # via
    #   django-fobi
    #   django-nonefield
django-nonefield==0.4
    # via django-fobi
django-phonenumber-field==5.1.0
    # via -r requirements.in
django-polymorphic-tree==2.0
    # via -r requirements.in
django-polymorphic==3.0.0
    # via django-polymorphic-tree
django-solo==1.1.5
    # via -r requirements.in
django-storages==1.11.1
    # via -r requirements.in
django-tables2==2.4.0
    # via -r requirements.in
django-tag-parser==3.2
    # via django-polymorphic-tree
django-treebeard==4.5.1
    # via -r requirements.in
django-vanilla-views==3.0.0
    # via -r requirements.in
django-widget-tweaks==1.4.8
    # via -r requirements.in
django==3.2.3
    # via
    #   -r requirements.in
    #   django-admin-sortable2
    #   django-appconf
    #   django-debug-toolbar
    #   django-dynamic-forms
    #   django-extensions
    #   django-extra-views
    #   django-formtools
    #   django-haystack
    #   django-mptt
    #   django-nine
    #   django-phonenumber-field
    #   django-polymorphic
    #   django-storages
    #   django-tables2
    #   django-treebeard
    #   djangorestframework
    #   easy-thumbnails
djangorestframework==3.12.4
    # via -r requirements.in
easy-thumbnails==2.7.1
    # via django-fobi
entrypoints==0.3
    # via -r requirements.in
enum34==1.1.10
    # via -r requirements.in
factory-boy==3.2.0
    # via -r requirements.in
faker==8.1.4
    # via factory-boy
funcsigs==1.0.2
    # via -r requirements.in
future==0.18.2
    # via -r requirements.in
google-auth-oauthlib==0.4.4
    # via gspread
google-auth==1.30.0
    # via
    #   google-auth-oauthlib
    #   gspread
gspread==3.7.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.19.1
    # via oauth2client
idna==2.10
    # via requests
ipaddress==1.0.23
    # via -r requirements.in
isort==5.8.0
    # via -r requirements.in
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
oauth2client==4.1.3
    # via -r requirements.in
oauthlib==3.1.0
    # via requests-oauthlib
olefile==0.46
    # via -r requirements.in
ordereddict==1.1
    # via -r requirements.in
packaging==20.9
    # via bleach
phonenumbers==8.12.23
    # via -r requirements.in
phonenumberslite==8.12.23
    # via -r requirements.in
pilkit==2.0
    # via -r requirements.in
pillow==8.2.0
    # via
    #   django-fobi
    #   easy-thumbnails
psycopg2==2.8.6
    # via -r requirements.in
purl==1.5
    # via -r requirements.in
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pycountry==20.7.3
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyopenssl==20.0.1
    # via -r requirements.in
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
python-dateutil==2.8.1
    # via
    #   botocore
    #   faker
pytz==2021.1
    # via
    #   babel
    #   django
rcssmin==1.0.6
    # via django-compressor
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   django-fobi
    #   requests-oauthlib
    #   rollbar
rjsmin==1.1.0
    # via django-compressor
rollbar==0.15.2
    # via -r requirements.in
rsa==4.7.2
    # via
    #   google-auth
    #   oauth2client
s3transfer==0.4.2
    # via boto3
selenium==3.141.0
    # via -r requirements.in
simplejson==3.17.2
    # via -r requirements.in
six==1.16.0
    # via
    #   bleach
    #   django-compressor
    #   django-extra-views
    #   django-fobi
    #   google-auth
    #   oauth2client
    #   purl
    #   pyopenssl
    #   python-dateutil
    #   rollbar
    #   vishap
sorl-thumbnail==12.7.0
    # via -r requirements.in
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via faker
unicodecsv==0.14.1
    # via -r requirements.in
unidecode==1.2.0
    # via django-fobi
uritemplate==3.0.1
    # via -r requirements.in
urllib3==1.26.4
    # via
    #   botocore
    #   requests
    #   selenium
vishap==0.1.5
    # via django-fobi
webencodings==0.5.1
    # via bleach
werkzeug==2.0.0
    # via -r requirements.in
whitenoise==5.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
